+++ client-app/shared/account/components/index.ts
@@ -10,4 +10,4 @@ export { default as ProfileUpdateSuccessDialog } from "./profile-update-success-
 export { default as OrdersFilter } from "./orders-filter.vue";
 export { default as MobileOrdersFilter } from "./mobile-orders-filter.vue";
 export { default as PageToolbarBlock } from "./page-toolbar-block.vue";
-export { default as CartLineItems } from "./cart-line-items.vue";
+export { default as QuoteLineItems } from "./quote-line-items.vue";
